more int types fix

+++ src/codegen.rs
@@ -90,7 +90,7 @@ pub fn lexer_struct(cx: &mut ExtCtxt, sp: Span, ident:Ident, props: &[Prop]) -> 
                 ast::Public
             ),
             id: -1 as u32,
-            ty: quote_ty!(&*cx, u64),
+            ty: quote_ty!(&*cx, usize),
             attrs: vec!()
         }
     });
@@ -172,7 +172,7 @@ fn simple_follow_method(cx:&mut ExtCtxt, sp:Span, lex:&Lexer) -> P<Method> {
     // * transtable: an array of N arrays of 256 uints, N being the number
     //   of states in the FSM, which gives the transitions between states
     let ty_vec = cx.ty(sp, ast::TyFixedLengthVec(
-        cx.ty_ident(sp, cx.ident_of("u64")),
+        cx.ty_ident(sp, cx.ident_of("usize")),
         cx.expr_uint(sp, 256)));
     let mut transtable = Vec::new();
 
@@ -195,7 +195,7 @@ fn simple_follow_method(cx:&mut ExtCtxt, sp:Span, lex:&Lexer) -> P<Method> {
             transtable);
 
     quote_method!(cx,
-        fn follow(&self, current_state:u64, symbol:u64) -> u64 {
+        fn follow(&self, current_state:usize, symbol:usize) -> usize {
             $transtable
             return TRANSITION_TABLE[current_state][symbol];
         }
@@ -206,7 +206,7 @@ fn simple_accepting_method(cx:&mut ExtCtxt, sp:Span, lex:&Lexer) -> P<Method> {
     // * accepting: an array of N uints, giving the action associated to
     //   each state
     let ty_acctable = cx.ty(sp, ast::TyFixedLengthVec(
-        cx.ty_ident(sp, cx.ident_of("u64")),
+        cx.ty_ident(sp, cx.ident_of("usize")),
         cx.expr_uint(sp, lex.auto.states.len())));
 
     let mut acctable = Vec::new();
@@ -220,7 +220,7 @@ fn simple_accepting_method(cx:&mut ExtCtxt, sp:Span, lex:&Lexer) -> P<Method> {
             acctable);
 
     quote_method!(cx,
-        fn accepting(&self, state:u64) -> u64 {
+        fn accepting(&self, state:usize) -> usize {
             $acctable
             return ACCEPTING[state];
         }
@@ -314,8 +314,8 @@ pub fn user_lexer_impl(cx: &mut ExtCtxt, sp: Span, lex:&Lexer) -> Vec<P<ast::Ite
                         _ => break
                     };
 
-                    let new_st:u64 = self.follow(current_st, i as u64);
-                    let action_id:u64 = self.accepting(new_st);
+                    let new_st:usize = self.follow(current_st, i as usize);
+                    let action_id:usize = self.accepting(new_st);
 
                     if action_id != 0 {
                         self._input.advance = self._input.pos;
